Don't ignore Markdown options in Haml filter
This issue was detailed in #1152.

Essentially we're not passing the :markdown options defined in config.rb to
the Markdown renderer when we first go through the Haml renderer.

The solution I've found is to fetch the appropriate options (not just for
Markdown but for Sass or SCSS also) within the Haml renderer's filter loop which
means this patch should also fix any ignored options for Sass and SCSS.