Update "Port in use"-message for PreviewServer (#2089)
When starting the preview server while the port is already in use, users
get the following error message:

  Port "4567" is in use. This should not have happened. Please start "middleman server" again.

I see two problems with this message:

1. "This should not have happened" makes this error sound like
   something scary went wrong, while this is something most middleman
   users will run into sooner or later.
2. "Please start "middleman server" again." isn't the best advise, since
   trying to start it again won't solve the issue, as there's either
   another middleman instance running, or some other process is keeping
   the port occupied.

Instead, I'd like to propose this:

  Port 4567 is already in use. This could mean another instance of middleman is already running. Please make sure port 4567 is free and start `middleman server` again, or choose another port by running `middleman server —-port=4568` instead.

This error message mentions that there might be another instance running
(which is quite likely in this situation), and tells the user to make
sure the port is free and try again *or* choose another port. The
suggested alternative port is `http_opts[:Port] + 1`, so it'll use 4568
if the current port is 4567.