Extending fixture data to fix tests
With the introduction of i18n fallbacks for localized paths, the tests which set the default locale to Spanish failed, since the Spanish path mappings were now used for the English locale as well. So the expected paths were now incorrect.

This change adds the expected paths to the English i18n file, so that the tests pass again.