Update magic hashes for feature spec
Fixes:
```
Then I should see 'href="../stylesheets/uses_partials-58447acf.css' # lib/middleman-core/step_definitions/server_steps.rb:94
  expected "<!DOCTYPE html>\n<html>\n  <head>\n    <meta charset=\"utf-8\">\n\n    <link href=\"../stylesheets/s...tylesheets/uses_partials-a3c8302b.css\" rel=\"stylesheet\" />\n\n    </div>\n\n  </body>\n</html>\n" to include "href=\"../stylesheets/uses_partials-58447acf.css"
  Diff:
  @@ -1,2 +1,20 @@
  -href="../stylesheets/uses_partials-58447acf.css
  +<!DOCTYPE html>
  +<html>
  +  <head>
  +    <meta charset="utf-8">
  +
  +    <link href="../stylesheets/site-47c40978.css" rel="stylesheet" /><link href="../stylesheets/fragment-313b648b.css" rel="stylesheet" />
  +    <script src="../javascripts/application-1d8d5276.js"></script>
  +
  +  </head>
  +
  +  <body class="partials partials_index">
  +
  +    <div id="main" role="main">
  +      <link href="../stylesheets/uses_partials-a3c8302b.css" rel="stylesheet" />
  +
  +    </div>
  +
  +  </body>
  +</html>
   (RSpec::Expectations::ExpectationNotMetError)
  ./lib/middleman-core/step_definitions/server_steps.rb:95:in `/^I should see '([^]*)'$/'
  features/asset_hash.feature:182:in `Then I should see 'href="../stylesheets/uses_partials-58447acf.css''
```

The `Scenario: The asset hash should change when a SASS partial changes`
test checks to make sure the hash of CSS file is updated when one of its
imported partial SASS files is changed. For some reason, the actual hash
values changed, but they are still different from each other before and
after the edit, so the feature still appears to be intact.  We also
tested this feature manually once to make sure there wasn't a
regression.

We also added a newline to the end of `partial.sass`. Partly because
this is what our IDE, RubyMine, does for us automatically, but also
because it's seen as standard practice in text files. We're hoping that
defaulting to a standard practice may cause less drift in this file, and
thus make it less likely that the hash changes.

Signed-off-by: Miranda Limonczenko <mlimonczenko@pivotal.io>