Reference contribution guidelines in README
Information is duplicated between the two. I think it is simpler to keep
the README focused on a Middleman user's perspective, whereas
CONTRIBUTING.md can focus on a contributor's perspective.

My motivation for this is that I'd like to later add more extensive
documentation about testing Middleman from a contributor's perspective,
but doing this in the README seemed like it would bloat it too much.