Fix linking of external urls when i18n is enabled (#2343)
The Internationalization extension is redefining url_for to localize url
paths. However, this redefinition was also acting on external urls,
causing them to sometimes have an "/index.html" appended to them.

This path uses the presence of a host in the parsed URI object to
determine whether or not a URL is "external". This is the same method
for determining this as is already used in the base url_for.