Move i18n test cleanup code to After block (#2345)
Avoiding having test specific code in the extension simplifies it, and
eliminates the possibility it would create a bug in a real Middleman
app.

Using `defined?(I18n)` is not strictly required, as the I18n module is
currently always present, as it is loaded via requiring
'active_support/all'. However, as that's not immediately apparent, and
could theoretically change in the future, I thought it was better to
have.