Ruby 3.0 support (#2423)
* Test against Ruby 3.0

* Add dependency on WEBrick in Ruby 3.0 and above

Ruby 3.0 removed WEBrick from the stdlib.

* Update Gemfile.lock's async-io for Ruby 3.0

Previous version was not compatible with Ruby 3.0

* pry-stack_explorer isn't Ruby 3.0 compatible

As it is dependent on binding_of_caller, which isn't Ruby 3.0
compatible, though there is an open PR
(https://github.com/banister/binding_of_caller/pull/80). As this Gem is
only included to make debugging in pry easier, avoiding automatically
requiring it seems like an acceptable compromise.