Run tests using GitHub workflows (#2425)
travis-ci.org now displays "Please be aware travis-ci.org will be
shutting down in several weeks, with all accounts migrating to
travis-ci.com. Please stay tuned here for more information."

As GitHub actions is free for public repositories, whereas it's unclear
to what travis-ci.com will be, and the switching cost is low, it seems
going with workflows is the prudent thing to do.