Fix tilt 2.5 deprecations
When using tilt 2.5, you get a lot of deprecation messages like this:

```
Lazy loading of org templates is deprecated and will be removed in Tilt 3.  Require org-ruby manually.
```

And similarly with many other templates.

This change seems to fix the problem but not eagerly looping through all
templates to remove the ones not available.

`:Tilt.default_mapping.lazy_map` is explicitly documented as private, so
not using that seems like a good change too.