Fix erb preview regression (#2781)
The ERB engine was switched to Erubi recently. The Erubi template
modifies template data in-place, when it has apostrophes. Since the data
we pass to the template holds a reference to the front matter file
cache, it will be modified for the subsequent request.

We need to pass a copy instead.