Workaround activesupport issue
The missing logger require is fixed in activesupport v8 or higher. And
middleman can be used with Rails 8 as long as Ruby 3.2 or higher is
used.

So add the missing logger dependency whenever activesupport is not
modern enough, and old Ruby is used.

Users of more modern rubies may still run into the issue, but in that
case the preferred fix is to upgrade activesupport.